requirements.txt
panda3d==1.10.6
Pillow==7.2.0
smmap==3.0.4
